module_4_vla/requirements.txt
# Core dependencies for VLA module
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
numpy>=1.21.0
opencv-python>=4.8.0
pydub>=0.25.1
sounddevice>=0.4.6

# ROS 2 related
rclpy>=3.0.0
ros2-message>=0.0.0

# Additional utilities
requests>=2.28.0
pyyaml>=6.0